NO-TICKET: Split contracts to separate workspace and provide build script
This commit adds a pseudo lib target to the contracts dir and a build
script for it which causes contracts to be compiled as Wasm binaries.
This allows contracts or subsets of them (e.g. system contracts) to be
added as dependencies of other crates.

Using this mechanism, the system contracts are now a dependency of
'casperlabs-engine-core' and all contracts are a dependency of
'casperlabs-engine-tests'.

This means that no tests need to be '#[ignore]d' as the required
contracts will always exist at the time of running the tests.  In this
way, we can now just do 'cargo test' in 'execution-engine' to run the
full suite.

The Makefile has been simplified, but retains most of the the previous
functionality.  'make build-contracts' has been changed to require a
specific subset of contracts (e.g. 'make build-contracts/system' builds
the system ones).

The execution-engine contracts have all been updated to specify that
tests, doctests and benchmarks should not be built, since attempting to
build these caused errors due to multiple definitions of standard lib
functions.  This now means that 'cargo clippy --all-targets' won't error
out.